% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sew.R
\name{sew.knitxl_output_text}
\alias{sew.knitxl_output_text}
\alias{sew.knitxl_output_vector}
\alias{sew.knitxl_output_data_frame}
\title{Extend \code{\link[knitr:sew]{knitr::sew()}} generic to write text, vectors and data table to the
output .xlsx file.}
\usage{
\method{sew}{knitxl_output_text}(x, options, ...)

\method{sew}{knitxl_output_vector}(x, options, ...)

\method{sew}{knitxl_output_data_frame}(x, options, ...)
}
\arguments{
\item{x}{Output from evaluate::evaluate().}

\item{options}{A list of chunk options used to control output.}

\item{...}{Other arguments to pass to methods.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
These functions are called within \code{\link[knitr:knit]{knitr::knit()}} to write either text
vector or a table into the output .xlsx file produced by \code{\link[=knitxl]{knitxl()}}.
}
\keyword{internal}
