% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_xl.R
\name{xl_renderer}
\alias{xl_renderer}
\title{Represents an R object into a format that can be printed into an XLSX file}
\usage{
xl_renderer(x, options)
}
\arguments{
\item{x}{the object to be rendered in the XLSX file.}

\item{options}{the `knitr` and `knitxl` options used to cutomize the
rendering of `x`}
}
\description{
This is a generic function that is intended for developers who want
to extend `knitxl` to print new classes R objects. It transforms an object
into a `knitxl_output_*` class (either `text`, `vector` or `data_frame`)
that can be printed in an XLSX file.
}
